Trait lea_mem::GcStrategy

pub trait GcStrategy {
    fn collect_step<C: GcCallback>(&mut self, &mut C);
    fn collect_atomic<C: GcCallback>(&mut self, &mut C);
    fn register_obj<T: Any>(&mut self, T) -> TracedRef<T>;
    fn intern_str<T: Into<Str>>(&mut self, T) -> TracedRef<Str>;

    unsafe fn get_ref<T: Any>(&self, t: TracedRef<T>) -> &T { ... }
    unsafe fn get_mut<T: Any>(&mut self, t: TracedRef<T>) -> &mut T { ... }
}

Trait for Garbage Collector implementations.

Required Methods

fn collect_step<C: GcCallback>(&mut self, &mut C)

Instruct the GC to perform a collection step. This can be called by user code. A callback is passed so the GC can run finalizers and find the roots.

fn collect_atomic<C: GcCallback>(&mut self, &mut C)

Perform an atomic collection. The GC should try to free all currently unreachable objects.

fn register_obj<T: Any>(&mut self, T) -> TracedRef<T>

Called when a garbage-collected object is created. The GC should take ownership of the object and can dispose of it when collecting.

fn intern_str<T: Into<Str>>(&mut self, T) -> TracedRef<Str>

Interns a string and returns a TracedRef to the shared instance.

Provided Methods

unsafe fn get_ref<T: Any>(&self, t: TracedRef<T>) -> &T

Given a TracedRef of an object owned by this GC, this method unsafely returns a reference to the object.

unsafe fn get_mut<T: Any>(&mut self, t: TracedRef<T>) -> &mut T

The mutable version of get_ref, returns a mutable reference to the object.

This will mutably borrow the GC and thus prevent collections and aliasing references (from get_ref).